Eliminated ElfLoader32.SectionHeaders

+++ src/ImageLoaders/Elf/RelSegmentRenderer.cs
@@ -54,7 +54,7 @@ namespace Reko.ImageLoaders.Elf
                     return;
 
                 uint sym = info >> 8;
-                string symStr = loader.GetSymbolName(symtab, (int)sym);
+                string symStr = loader.GetSymbolName(symtab, sym);
                 formatter.Write("{0:X8} {1,3} {2:X8} {3}", offset, info & 0xFF, sym, symStr);
                 formatter.WriteLine();
             }
